Changed "Invalid data + Valid container" behavior
v1 returned `false` if context is invalid, even if the container is
valid. v2 will now return the element but with undefined data instead
(so the application can decide is the missing context defaults to
fallback data or if it's a fatal error instead of getContext assuming
it's a fatal error).